# docs/openapi-sprint1.json
{"openapi":"3.1.0","info":{"title":"NeoBank Platform API","description":"REST API for NeoBank — PMIS Internship · Infosys Bhubaneswar DC","contact":{"name":"NeoBank Dev Team","email":"dev@neobank.in"},"version":"Sprint 1 · v1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"Bearer Auth":[]}],"tags":[{"name":"Authentication","description":"Register and login endpoints"},{"name":"Accounts","description":"Bank account management"},{"name":"Transactions","description":"Transaction history and creation"}],"paths":{"/api/users/me":{"get":{"tags":["user-controller"],"operationId":"getProfile","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}},"put":{"tags":["user-controller"],"operationId":"updateProfile","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProfileRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}}},"/api/auth/register":{"post":{"tags":["Authentication"],"summary":"Register new user","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"User created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"409":{"description":"Email exists","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}},"400":{"description":"Validation error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RegisterResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["Authentication"],"summary":"Login and get JWT token","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Login success","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"401":{"description":"Bad credentials","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"403":{"description":"Account inactive","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}}}}},"/api/accounts":{"get":{"tags":["Accounts"],"summary":"Get all accounts owned by current user","operationId":"getMyAccounts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AccountResponse"}}}}}},"security":[{"Bearer Auth":[]}]},"post":{"tags":["Accounts"],"summary":"Create a new bank account","operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccountRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}},"security":[{"Bearer Auth":[]}]}},"/api/accounts/{accountId}/transactions":{"get":{"tags":["Transactions"],"summary":"Get paginated transaction history","operationId":"getTransactions","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTransactionResponse"}}}}},"security":[{"Bearer Auth":[]}]},"post":{"tags":["Transactions"],"summary":"Create debit or credit transaction","operationId":"createTransaction","parameters":[{"name":"accountId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransactionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionResponse"}}}}},"security":[{"Bearer Auth":[]}]}},"/api/admin/users":{"get":{"tags":["admin-controller"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserProfileResponse"}}}}}}}},"/api/accounts/{id}":{"get":{"tags":["Accounts"],"summary":"Get account by ID","operationId":"getAccountById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AccountResponse"}}}}},"security":[{"Bearer Auth":[]}]}}},"components":{"schemas":{"UpdateProfileRequest":{"type":"object","properties":{"fullName":{"type":"string"}},"required":["fullName"]},"UserProfileResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"isActive":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"RegisterRequest":{"type":"object","properties":{"fullName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string","maxLength":2147483647,"minLength":8,"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&#])[A-Za-z\\d@$!%*?&#]{8,}$"}},"required":["email","fullName","password"]},"RegisterResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fullName":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"isActive":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}},"LoginRequest":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponse":{"type":"object","properties":{"token":{"type":"string"},"email":{"type":"string"},"role":{"type":"string"},"userId":{"type":"integer","format":"int64"}}},"AccountRequest":{"type":"object","properties":{"accountType":{"type":"string","enum":["SAVINGS","CURRENT"]}},"required":["accountType"]},"AccountResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"accountNumber":{"type":"string"},"balance":{"type":"number"},"accountType":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"TransactionRequest":{"type":"object","properties":{"type":{"type":"string","enum":["DEBIT","CREDIT"]},"amount":{"type":"number","minimum":0.01},"description":{"type":"string","maxLength":500,"minLength":0}},"required":["amount","type"]},"TransactionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"type":{"type":"string"},"amount":{"type":"number"},"description":{"type":"string"},"transactionDate":{"type":"string","format":"date-time"},"balanceAfter":{"type":"number"}}},"PageTransactionResponse":{"type":"object","properties":{"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"},"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"SortObject":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}},"securitySchemes":{"Bearer Auth":{"type":"http","name":"Bearer Auth","scheme":"bearer","bearerFormat":"JWT"}}}}