# src/main/java/com/neobank/backend/account/AccountController.java
package com.neobank.backend.account;

import com.neobank.backend.account.dto.*;
import jakarta.validation.Valid;
import lombok.RequiredArgsConstructor;
import org.springframework.http.*;
import org.springframework.security.core.annotation.AuthenticationPrincipal;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import io.swagger.v3.oas.annotations.*;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;

@Tag(name = "Accounts", description = "Bank account management")
@RestController
@RequestMapping("/api/accounts")
@RequiredArgsConstructor
public class AccountController {

    private final AccountService accountService;

    @Operation(summary = "Create a new bank account",
               security = @SecurityRequirement(name = "Bearer Auth"))
    @PostMapping
    public ResponseEntity<AccountResponse> createAccount(
            @AuthenticationPrincipal UserDetails userDetails,
            @Valid @RequestBody AccountRequest request) {

        AccountResponse response = accountService
                .createAccount(userDetails.getUsername(), request);
        return ResponseEntity.status(HttpStatus.CREATED).body(response);
    }

    @Operation(summary = "Get all accounts owned by current user",
               security = @SecurityRequirement(name = "Bearer Auth"))
    @GetMapping
    public ResponseEntity<List<AccountResponse>> getMyAccounts(
            @AuthenticationPrincipal UserDetails userDetails) {

        return ResponseEntity.ok(
            accountService.getAccountsByUser(userDetails.getUsername()));
    }

    @Operation(summary = "Get account by ID",
               security = @SecurityRequirement(name = "Bearer Auth"))
    @GetMapping("/{id}")
    public ResponseEntity<AccountResponse> getAccountById(
            @AuthenticationPrincipal UserDetails userDetails,
            @PathVariable Long id) {

        return ResponseEntity.ok(
            accountService.getAccountById(id, userDetails.getUsername()));
    }
}